+++ src/constants/rules.js
@@ -1,0 +1,16 @@
+export const RULES = [
+  '+',
+  '-',
+  '/',
+  '*',
+  '1',
+  '2',
+  '3',
+  '4',
+  '5',
+  '6',
+  '7',
+  '8',
+  '9',
+  '0',
+];
