+++ src/screens/Display/index.js
@@ -3,16 +3,12 @@ import {View, Text, TextInput, Alert} from 'react-native';
 
 import CalcButton from '../../components/CalcButtons';
 import {styles} from './style';
+import {createButtons} from '../../utils/createButtons';
+import {RULES} from '../../constants/rules';
 
 const Display = () => {
   const [userText, setUserText] = useState('');
   const [calcText, setCalcText] = useState('');
-  const handlePress = number => {
-    if (number === '=') {
-      return calculation();
-    }
-    setUserText(userText + number);
-  };
 
   const calculation = () => {
     // eslint-disable-next-line no-eval
@@ -20,22 +16,6 @@ const Display = () => {
   };
 
   const handleInput = event => {
-    const RULES = [
-      '+',
-      '-',
-      '/',
-      '*',
-      '1',
-      '2',
-      '3',
-      '4',
-      '5',
-      '6',
-      '7',
-      '8',
-      '9',
-      '0',
-    ];
     if (event.nativeEvent.key === 'Backspace') {
       setUserText(userText.toString().substring(0, userText.length - 1));
       return;
@@ -54,55 +34,7 @@ const Display = () => {
     );
   };
 
-  const handleOperation = op => {
-    if (op === 'Clear') {
-      setUserText('');
-      setCalcText(0);
-      return;
-    }
-
-    if (op === 'Del') {
-      setUserText(userText.toString().substring(0, userText.length - 1));
-      return;
-    }
-
-    if (
-      ['Del', 'Clear', '+', '-', '*', '/'].includes(
-        userText.toString().split('').pop(),
-      )
-    ) {
-      return;
-    }
-    setUserText(userText + op);
-  };
-
-  const BUTTONS = [
-    {
-      type: 'operations',
-      labels: ['Del', 'Clear', '+', '-', '*', '/'],
-      action: b => handleOperation(b),
-    },
-    {
-      type: 'numbers',
-      labels: [7, 8, 9],
-      action: b => handlePress(b),
-    },
-    {
-      type: 'numbers',
-      labels: [4, 5, 6],
-      action: b => handlePress(b),
-    },
-    {
-      type: 'numbers',
-      labels: [1, 2, 3],
-      action: b => handlePress(b),
-    },
-    {
-      type: 'numbers',
-      labels: ['.', 0, '='],
-      action: b => handlePress(b),
-    },
-  ];
+  const buttons = createButtons(userText, setUserText, setCalcText);
 
   return (
     <View style={styles.container}>
@@ -117,7 +49,7 @@ const Display = () => {
           value={userText}
         />
       </View>
-      <CalcButton buttons={BUTTONS} />
+      <CalcButton buttons={buttons} />
     </View>
   );
 };
